Further enhancements in annotation and documentation, notably in Config. Fixed Trie enumerators throwing ArgumentNullException when trying to enumerate an empty trie/subtrie. Color.Parse/Color.GetName/Rank.Parse no longer accept nulls.

+++ branch-0.90x/fCraft/Commands/MaintenanceCommands.cs
@@ -1293,8 +1293,10 @@ namespace fCraft {
 
 
         static void PruneDBTask( [NotNull] SchedulerTask task ) {
-            int removedCount = PlayerDB.RemoveInactivePlayers();
             Player player = (Player)task.UserState;
+            if( player == null ) throw new NullReferenceException( "task.UserState" );
+
+            int removedCount = PlayerDB.RemoveInactivePlayers();
             player.Message( "PruneDB: Removed {0} inactive players!", removedCount );
         }
